Chatbot_Kemendukbangga/requirements.txt
pandas
numpy
sentence-transformers
faiss-cpu
scikit-learn
rapidfuzz
whisper
soundfile
pydub
flask
gunicorn
uvicorn
python-dotenv
google-generativeai